+++ lib/core/firebase/types.dart
@@ -1,2 +1,2 @@
-export 'types/auth_response.dart';
 export 'types/backup_document.dart';
+export 'types/response.dart';