+++ lib/modules/settings/routes/account/widgets.dart
@@ -1,3 +1,4 @@
-export 'widgets/backup_tile.dart';
-export 'widgets/delete_tile.dart';
-export 'widgets/restore_tile.dart';
+export 'widgets/config_backup_tile.dart';
+export 'widgets/config_delete_tile.dart';
+export 'widgets/config_restore_tile.dart';
+export 'widgets/delete_account_tile.dart';